Orders integration tests

+++ commercetools.Sdk/commercetools.Sdk.Domain/Orders/ReturnItemDraft.cs
@@ -2,7 +2,7 @@
 {
     public abstract class ReturnItemDraft
     {
-        public double Quantity { get; set; }
+        public long Quantity { get; set; }
         public string Comment { get; set; }
         public ReturnShipmentState ShipmentState { get; set; }
     }
